[refactor] split serialization_utils into v7/v8 part files with explicit version naming
